moving all the different stimuli classes into seperate modules and trying to sort them
