+++ rewardgym/stimuli/create_images.py
